requirements.txt
gitpython
openai
python-dotenv
tqdm
requests
networkx
matplotlib
pydot
networkx